backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
Pillow
pywhatkit
openpyxl
pywin32
pyautogui
